+++ src/langserver/symbolsinformation/toolsGetDependenciesListRequest.ts
@@ -1,0 +1,7 @@
+export class ToolsGetDependenciesListRequest {
+    path: string;
+
+    constructor(newPath: string) {
+        this.path = newPath;
+    }
+}