+++ src/langserver/symbolsinformation/toolsGetPageDetailsRequest.ts
@@ -2,10 +2,12 @@ import { ToolsGetTableBasedSymbolDetailsRequest } from "./toolsGetTableBasedSymb
 
 export class ToolsGetPageDetailsRequest  extends ToolsGetTableBasedSymbolDetailsRequest {
     getToolTips: boolean;
+    toolTipsSourceDependencies: string[] | undefined;
 
 
-    constructor(newPath: string | undefined, newName: string, newGetExistingFields: boolean, newGetAvailableFields: boolean, newGetToolTips: boolean) {
+    constructor(newPath: string | undefined, newName: string, newGetExistingFields: boolean, newGetAvailableFields: boolean, newGetToolTips: boolean, newToolTipsSourceDependencies: string[] | undefined) {
         super(newPath, newName, newGetExistingFields, newGetAvailableFields);
         this.getToolTips = newGetToolTips;
+        this.toolTipsSourceDependencies = newToolTipsSourceDependencies;
     }
 }