+++ src/langserver/symbolsinformation/toolsGetPageFieldAvailableToolTipsRequest.ts
@@ -1,0 +1,16 @@
+export class ToolsGetPageFieldAvailableToolTipsRequest {
+    path: string;
+    objectType: string;
+    objectName: string;
+    sourceTable: string;
+    fieldExpression: string;
+
+    constructor(newPath: string, newObjectType: string, newObjectName: string, newSourceTable: string, newFieldExpression: string) {
+        this.path = newPath;
+        this.objectType = newObjectType;
+        this.objectName = newObjectName;
+        this.sourceTable = newSourceTable;
+        this.fieldExpression = newFieldExpression;
+    }
+
+}