+++ src/objectwizards/wizards/alPageWizardPage.ts
@@ -14,7 +14,8 @@ export class ALPageWizardPage extends ALTableBasedWizardPage {
     constructor(toolsExtensionContext : DevToolsExtensionContext, settings: ALObjectWizardSettings, data : ALPageWizardData) {
         super(toolsExtensionContext, "AL Page Wizard", settings, data);
         this._pageWizardData = data;
-        this._includeToolTips = true;
+        this._includeToolTips = data.reuseToolTips;
+        this._toolTipsSourceDependencies = data.toolTipsSource;
     }
 
     protected getHtmlContentPath() : string {